bash/cpu: fix resolution of symlinked sshfs mounts
If you mount your remote sshfs(1) file systems under /Volumes and
symlink it into your home directory, cpu.sh had a bug where it did
not resolve the symlink.  This caused CPU_REMOTE to be set to the
full path of the symlink instead of the relative path of the remote
directory.